requirements.txt
# Django Core
Django==5.0.6
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Environment Variables
python-dotenv==1.0.1

# AI & LangChain
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-community==0.2.16
langchain-text-splitters==0.2.4

# Google Generative AI
google-generativeai==0.7.2

# Vector Database
faiss-cpu==1.8.0

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.0

# Embeddings (open-source fallback)
sentence-transformers==3.0.1

# Utilities
Pillow==10.4.0
whitenoise==6.7.0
